Player count made into it's own component.
